UCS/TOPO: Move VR files into its own folder
