+++ src/ucs/Makefile.am
@@ -134,7 +134,7 @@ noinst_HEADERS = \
 	sys/module_int.h \
 	sys/sys.h \
 	sys/topo/base/topo_int.h \
-	sys/topo/vr.h \
+	sys/topo/vr/vr.h \
 	sys/iovec.h \
 	sys/iovec.inl \
 	sys/ptr_arith.h \
@@ -215,7 +215,7 @@ libucs_la_SOURCES = \
 	sys/lib.c \
 	sys/sock.c \
 	sys/topo/base/topo.c \
-	sys/topo/vr.c \
+	sys/topo/vr/vr.c \
 	sys/stubs.c \
 	sys/netlink.c \
 	sys/uid.c \
