+++ src/ucs/sys/topo/base/topo.c
@@ -9,7 +9,7 @@
 #endif
 
 #include "topo_int.h"
-#include <ucs/sys/topo/vr.h>
+#include <ucs/sys/topo/vr/vr.h>
 
 #include <ucs/arch/cpu.h>
 #include <ucs/memory/numa.h>
